On my 24 proc box I got ENOSPC, so make the routine return the size it wants and try again.

+++ sys/miscfs/procfs/procfs.h
@@ -1,4 +1,4 @@
-/*	$NetBSD: procfs.h,v 1.68 2012/05/28 13:16:10 christos Exp $	*/
+/*	$NetBSD: procfs.h,v 1.69 2014/04/05 18:42:32 christos Exp $	*/
 
 /*
  * Copyright (c) 1993
@@ -245,7 +245,7 @@ int procfs_validmap(struct lwp *, struct mount *);
 
 int procfs_rw(void *);
 
-int procfs_getcpuinfstr(char *, int *);
+int procfs_getcpuinfstr(char *, size_t *);
 
 #define PROCFS_LOCKED	0x01
 #define PROCFS_WANT	0x02
